Add files via upload
This file acts as one big query to create a database "carparks" and creates and populates the tables "carpark" and "space". Also a placeholder user table but it only has one column.
Hopefully you guys who also have made databases/tables can just add to the query so that we all get on the same page.